% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{*.molecule_single}
\alias{*.molecule_single}
\alias{/.molecule_single}
\alias{+.molecule_single}
\alias{combine_molecules}
\alias{==.molecule_single}
\alias{*.mol}
\alias{/.mol}
\alias{+.mol}
\alias{==.mol}
\title{Molecule arithmetic}
\usage{
\method{*}{molecule_single}(x, y)

\method{/}{molecule_single}(x, y)

\method{+}{molecule_single}(x, y)

combine_molecules(...)

\method{==}{molecule_single}(x, y)

\method{*}{mol}(x, y)

\method{/}{mol}(x, y)

\method{+}{mol}(x, y)

\method{==}{mol}(x, y)
}
\arguments{
\item{x}{A mol object}

\item{y}{The object with which to perform arithmetic}

\item{...}{Objects to combine}
}
\value{
A mol object
}
\description{
Molecule arithmetic
}
\examples{
m <- as_mol("H2O")
m*2
m + as_mol("H+")

}
