% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{as.character.reaction}
\alias{as.character.reaction}
\alias{print.reaction}
\title{Coerce reactions to a character vector}
\usage{
\method{as.character}{reaction}(x, ...)

\method{print}{reaction}(x, ...)
}
\arguments{
\item{x}{A reaction object}

\item{...}{Ignored}
}
\value{
A character vector
}
\description{
Coerce reactions to a character vector
}
\examples{
r <- as_reaction(2*H2 + O2 ~ 2*H2O)
as.character(r)
print(r)

}
