% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{as_markdown}
\alias{as_markdown}
\alias{as_markdown.default}
\alias{as_markdown.reaction}
\alias{as_markdown.reaction_list}
\title{Format objects as markdown}
\usage{
as_markdown(x, ...)

\method{as_markdown}{default}(x, ...)

\method{as_markdown}{reaction}(x, equals_sign = NULL, ...)

\method{as_markdown}{reaction_list}(x, ...)
}
\arguments{
\item{x}{A mol, reaction, or reaction list object}

\item{...}{Passed to as.character()}

\item{equals_sign}{An equals sign to use for a reaction object}
}
\value{
A character vector formatted for markdown
}
\description{
Format objects as markdown
}
\examples{
as_markdown(as_reaction("2H+ + 2OH- = 2H2O"))
as_markdown(as_reaction_list("2H+ + 2OH- = 2H2O"))
as_markdown(as_mol(c("SO4-2", "Ca+2")))

}
