% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{chem_expression}
\alias{chem_expression}
\alias{chem_expression.default}
\alias{chem_expression.reaction}
\alias{chem_expression.reaction_list}
\title{Format objects as plotmath expressions}
\usage{
chem_expression(x, ...)

\method{chem_expression}{default}(x, ...)

\method{chem_expression}{reaction}(x, equals_sign = NULL, ...)

\method{chem_expression}{reaction_list}(x, ...)
}
\arguments{
\item{x}{A mol, reaction, or reaction list object}

\item{...}{Passed to as.character()}

\item{equals_sign}{An equals sign to use for a reaction object}
}
\description{
For use on plots using parse = TRUE
}
\examples{
chem_expression(as_reaction("2H+ + 2OH- = 2H2O"))
chem_expression(as_reaction_list("2H+ + 2OH- = 2H2O"))
chem_expression(as_mol(c("SO4-2", "Ca+2")))

}
