% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\name{is_element}
\alias{is_element}
\alias{elsymbol}
\alias{elmass}
\alias{elz}
\alias{elname}
\alias{elgroup}
\alias{elperiod}
\alias{elvalence}
\title{Element properties}
\usage{
is_element(x)

elsymbol(z)

elmass(x)

elz(x)

elname(x, locale = "en_UK")

elgroup(x)

elperiod(x)

elvalence(x)
}
\arguments{
\item{x}{An element symbol (e.g. H, He)}

\item{z}{The atomic number Z}

\item{locale}{Currently not supported, but will eventually return "Al" as "Aluminum"
for unsavory Americans.}
}
\value{
A named vector of the relevant value(s)
}
\description{
Based on the periodic tibble (see \link{pt}).
}
\examples{
is_element("H")
is_element("Hg")
is_element("H2O")

elsymbol(1)

elmass("H")
elmass("O")

elz("Pb")
elname("Al")
elgroup("Al")
elperiod("Al")
elvalence("Al")

}
