% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\docType{data}
\name{elsymbols}
\alias{elsymbols}
\alias{elmasses}
\alias{elzs}
\alias{elnames}
\alias{elgroups}
\alias{elperiods}
\alias{elvalences}
\title{Element properties as lists}
\format{An object of class \code{list} of length 118.}
\usage{
elsymbols

elmasses

elzs

elnames

elgroups

elperiods

elvalences
}
\description{
Based on the periodic tibble (see \link{pt}).
}
\examples{
with(elmasses, 2*H + O)
with(elzs, H:O)

}
\keyword{datasets}
