% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{is_balanced}
\alias{is_balanced}
\alias{balance}
\title{Balance reactions}
\usage{
is_balanced(x, charge = TRUE, tol = .Machine$double.eps^0.5)

balance(x, charge = TRUE, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A reaction object}

\item{charge}{Balance charge?}

\item{tol}{Tolerance for zero detection}
}
\value{
A balanced reaction or logical describing if the reaction is balanced
}
\description{
Balance reactions
}
\examples{
is_balanced("O2 + 2H2 = 2H2O")
is_balanced("O2 + H2 = 2H2O")
is_balanced("O2-4 + 2H2 = 2H2O", charge = FALSE)
is_balanced("O2-4 + 2H2 = 2H2O", charge = TRUE)

}
