% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{mass}
\alias{mass}
\alias{mass.default}
\alias{mass.molecule_single}
\alias{mass.molecules}
\alias{charge}
\alias{charge.default}
\alias{charge.molecule_single}
\alias{charge.molecules}
\title{Access properties of a molecule object}
\usage{
mass(x)

\method{mass}{default}(x)

\method{mass}{molecule_single}(x)

\method{mass}{molecules}(x)

charge(x)

\method{charge}{default}(x)

\method{charge}{molecule_single}(x)

\method{charge}{molecules}(x)
}
\arguments{
\item{x}{A molecule object or symbol string}
}
\value{
The mass of the molecule or element in g/mol
}
\description{
Access properties of a molecule object
}
\examples{
mass("H")
m <- as_molecule_single("H2O")
mass(m)

}
