% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{data}
\name{mol}
\alias{mol}
\alias{molecule_single}
\alias{as_molecule_single}
\alias{as_mol}
\alias{as.mol}
\alias{as_molecule_single.molecule_single}
\alias{as_molecule_single.mol}
\alias{as_molecule_single.character}
\alias{as_molecule_single.formula}
\alias{as_mol.mol}
\alias{as_mol.molecule_single}
\alias{as_mol.character}
\alias{as_mol.formula}
\alias{new_mol}
\alias{NA_molecule_}
\alias{validate_molecule_single}
\alias{validate_mol}
\alias{is_molecule_single}
\alias{is_mol}
\title{Create a molecule object}
\format{An object of class \code{molecule_single} of length 0.}
\usage{
mol(..., validate = TRUE)

molecule_single(..., charge = 0L, validate = TRUE)

as_molecule_single(x, ...)

as_mol(x, ...)

as.mol(x, ...)

\method{as_molecule_single}{molecule_single}(x, ...)

\method{as_molecule_single}{mol}(x, ...)

\method{as_molecule_single}{character}(x, validate = TRUE, ...)

\method{as_molecule_single}{formula}(x, validate = TRUE, ...)

\method{as_mol}{mol}(x, ...)

\method{as_mol}{molecule_single}(x, validate = TRUE, ...)

\method{as_mol}{character}(x, validate = TRUE, ...)

\method{as_mol}{formula}(x, validate = TRUE, ...)

new_mol(x)

NA_molecule_

validate_molecule_single(x)

validate_mol(x)

is_molecule_single(x)

is_mol(x)
}
\arguments{
\item{...}{A named list of symbols and counts}

\item{validate}{Validate elements in molecule}

\item{charge}{The charge of the molecule}

\item{x}{An objec to be coerced to type molecule(s)}
}
\value{
An object of class molecule
}
\description{
Create a molecule object
}
\examples{
m1 <- molecule_single(H=2, O=1)
m2 <- molecule_single(H=1, charge = 1)
mol(m1, m2)

}
\keyword{datasets}
