% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{data}
\name{molecule}
\alias{molecule}
\alias{molecules}
\alias{as_molecule}
\alias{as_molecules}
\alias{as_molecule.molecule}
\alias{as_molecule.molecules}
\alias{as_molecule.character}
\alias{as_molecule.formula}
\alias{as_molecules.molecules}
\alias{as_molecules.molecule}
\alias{as_molecules.character}
\alias{as_molecules.formula}
\alias{new_molecules}
\alias{NA_molecule_}
\alias{validate_molecule}
\alias{validate_molecules}
\alias{is_molecule}
\alias{is_molecules}
\title{Create a molecule object}
\format{An object of class \code{molecule} of length 0.}
\usage{
molecule(..., charge = 0L, validate = TRUE)

molecules(..., validate = TRUE)

as_molecule(x, ...)

as_molecules(x, ...)

\method{as_molecule}{molecule}(x, ...)

\method{as_molecule}{molecules}(x, ...)

\method{as_molecule}{character}(x, validate = TRUE, ...)

\method{as_molecule}{formula}(x, validate = TRUE, ...)

\method{as_molecules}{molecules}(x, ...)

\method{as_molecules}{molecule}(x, validate = TRUE, ...)

\method{as_molecules}{character}(x, validate = TRUE, ...)

\method{as_molecules}{formula}(x, validate = TRUE, ...)

new_molecules(x)

NA_molecule_

validate_molecule(x)

validate_molecules(x)

is_molecule(x)

is_molecules(x)
}
\arguments{
\item{...}{A named list of symbols and counts}

\item{charge}{The charge of the molecule}

\item{validate}{Validate elements in molecule}

\item{x}{An objec to be coerced to type molecule(s)}
}
\value{
An object of class molecule
}
\description{
Create a molecule object
}
\examples{
m1 <- molecule(H=2, O=1)
m2 <- molecule(H=1, charge = 1)
molecules(m1, m2)

}
\keyword{datasets}
