% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{as.data.frame.molecule_single}
\alias{as.data.frame.molecule_single}
\alias{as_tibble.molecule_single}
\alias{as_tibble.mol}
\alias{as.data.frame.mol}
\alias{as.matrix.molecule_single}
\alias{as.matrix.mol}
\title{Data frame reprsentation of a molecule object}
\usage{
\method{as.data.frame}{molecule_single}(x, ...)

\method{as_tibble}{molecule_single}(x, ...)

\method{as_tibble}{mol}(x, ...)

\method{as.data.frame}{mol}(x, ...)

\method{as.matrix}{molecule_single}(x, ...)

\method{as.matrix}{mol}(x, ...)
}
\arguments{
\item{x}{A molecule object}

\item{...}{Ignored}
}
\value{
A tibble with one row per molecule
}
\description{
Data frame reprsentation of a molecule object
}
\examples{
as.data.frame(as_mol(~H2O))
library(tibble)
as_tibble(as_mol(~H2O))
as_tibble(mol("H2O", "NH3"))

}
