% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\docType{data}
\name{molecules}
\alias{molecules}
\alias{molecule_single}
\alias{as_molecule_single}
\alias{as_molecules}
\alias{as_molecule_single.molecule_single}
\alias{as_molecule_single.molecules}
\alias{as_molecule_single.character}
\alias{as_molecule_single.formula}
\alias{as_molecules.molecules}
\alias{as_molecules.molecule_single}
\alias{as_molecules.character}
\alias{as_molecules.formula}
\alias{new_molecules}
\alias{NA_molecule_}
\alias{validate_molecule_single}
\alias{validate_molecules}
\alias{is_molecule_single}
\alias{is_molecules}
\title{Create a molecule object}
\format{An object of class \code{molecule_single} of length 0.}
\usage{
molecules(..., validate = TRUE)

molecule_single(..., charge = 0L, validate = TRUE)

as_molecule_single(x, ...)

as_molecules(x, ...)

\method{as_molecule_single}{molecule_single}(x, ...)

\method{as_molecule_single}{molecules}(x, ...)

\method{as_molecule_single}{character}(x, validate = TRUE, ...)

\method{as_molecule_single}{formula}(x, validate = TRUE, ...)

\method{as_molecules}{molecules}(x, ...)

\method{as_molecules}{molecule_single}(x, validate = TRUE, ...)

\method{as_molecules}{character}(x, validate = TRUE, ...)

\method{as_molecules}{formula}(x, validate = TRUE, ...)

new_molecules(x)

NA_molecule_

validate_molecule_single(x)

validate_molecules(x)

is_molecule_single(x)

is_molecules(x)
}
\arguments{
\item{...}{A named list of symbols and counts}

\item{validate}{Validate elements in molecule}

\item{charge}{The charge of the molecule}

\item{x}{An objec to be coerced to type molecule(s)}
}
\value{
An object of class molecule
}
\description{
Create a molecule object
}
\examples{
m1 <- molecule_single(H=2, O=1)
m2 <- molecule_single(H=1, charge = 1)
molecules(m1, m2)

}
\keyword{datasets}
