% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{new_reaction}
\alias{new_reaction}
\alias{validate_reaction}
\alias{is_reaction}
\title{Create, validate reaction objects}
\usage{
new_reaction(x)

validate_reaction(x)

is_reaction(x)
}
\arguments{
\item{x}{A reaction object}
}
\value{
A reaction object
}
\description{
Create, validate reaction objects
}
\examples{
r <- new_reaction(list(lhs = as_mol(c("H2O", "H+")), rhs = as_mol("H3O+"),
                       counts_lhs = c(1, 1), counts_rhs = 1))
validate_reaction(r)

}
