% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{reaction}
\alias{reaction}
\alias{as_reaction}
\alias{as.reaction}
\alias{as_reaction.reaction}
\alias{as_reaction.formula}
\alias{as_reaction.character}
\title{Create reaction objects}
\usage{
reaction(lhs, rhs, counts_lhs = 1, counts_rhs = 1, validate = TRUE)

as_reaction(x, ...)

as.reaction(x, ...)

\method{as_reaction}{reaction}(x, ...)

\method{as_reaction}{formula}(x, validate = TRUE, ...)

\method{as_reaction}{character}(x, validate = TRUE, ...)
}
\arguments{
\item{lhs}{The lefthand side}

\item{rhs}{The righthand side}

\item{counts_lhs}{The counts for molecules on the lefthand side}

\item{counts_rhs}{The counts for molecules on the righthand side}

\item{validate}{Flag to validate molecules in the reaction}

\item{x}{An object to convert to a reaction object}

\item{...}{Passed to/from methods}
}
\value{
A reaction object
}
\description{
Create reaction objects
}
\examples{
reaction(~H2O + `H+`, ~`H3O+`)
reaction(c("H2O", "H+"), "H3O+")

}
