% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{as.data.frame.reaction}
\alias{as.data.frame.reaction}
\alias{as_tibble.reaction}
\alias{as.matrix.reaction}
\title{Data frame reprsentation of a reaction object}
\usage{
\method{as.data.frame}{reaction}(x, ...)

\method{as_tibble}{reaction}(x, ...)

\method{as.matrix}{reaction}(x, ...)
}
\arguments{
\item{x}{A reaction object}

\item{...}{Ignored}
}
\value{
A tibble with one row per molecule in reaction
}
\description{
Data frame reprsentation of a reaction object
}
\examples{
as.data.frame(as_reaction(O2 + 2*H2 ~ 2*H2O))
library(tibble)
as_tibble(as_reaction(O2 + 2*H2 ~ 2*H2O))

as.matrix(as_reaction(O2 + 2*H2 ~ 2*H2O))

}
