% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{[.reaction}
\alias{[.reaction}
\alias{lhs}
\alias{rhs}
\title{Subset, combine reaction objects}
\usage{
\method{[}{reaction}(x, i, ...)

lhs(x)

rhs(x)
}
\arguments{
\item{x}{A reaction object}

\item{i}{The index}

\item{...}{Ignored}
}
\value{
A reaction object
}
\description{
Subset, combine reaction objects
}
\examples{
r <- as_reaction("2H2 + O2 = 2H2O")
lhs(r)
rhs(r)

}
