% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{simplify}
\alias{simplify}
\alias{remove_zero_counts}
\alias{simplify.molecule_single}
\alias{remove_zero_counts.molecule_single}
\alias{simplify.mol}
\alias{remove_zero_counts.mol}
\title{Simplify chemr objects}
\usage{
simplify(x, ...)

remove_zero_counts(x, ...)

\method{simplify}{molecule_single}(x, ...)

\method{remove_zero_counts}{molecule_single}(x, tol = .Machine$double.eps^0.5,
  ...)

\method{simplify}{mol}(x, ...)

\method{remove_zero_counts}{mol}(x, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{A mol, molecule_single, or reaction object}

\item{...}{Unused}

\item{tol}{Tolerance for zero detection}
}
\value{
An object with the same class as x
}
\description{
Simplify chemr objects
}
\examples{
simplify(as_mol("CHOOOH"))

}
