% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{simplify_mol}
\alias{simplify_mol}
\alias{remove_zero_counts}
\alias{simplify_mol.molecule_single}
\alias{remove_zero_counts.molecule_single}
\alias{simplify_mol.mol}
\alias{remove_zero_counts.mol}
\title{Simplify chemr objects}
\usage{
simplify_mol(x, ...)

remove_zero_counts(x, ...)

\method{simplify_mol}{molecule_single}(x, ...)

\method{remove_zero_counts}{molecule_single}(x,
  tol = .Machine$double.eps^0.5, ...)

\method{simplify_mol}{mol}(x, ...)

\method{remove_zero_counts}{mol}(x, tol = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{A mol, molecule_single, or reaction object}

\item{...}{Unused}

\item{tol}{Tolerance for zero detection}
}
\value{
An object with the same class as x
}
\description{
Simplify chemr objects
}
\examples{
simplify_mol(as_mol("CHOOOH"))

}
