% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{simplify_reaction}
\alias{simplify_reaction}
\alias{simplify_reaction.reaction}
\alias{remove_zero_counts.reaction}
\alias{remove_zero_counts.reaction_list}
\alias{simplify_reaction.reaction_list}
\title{Simplify reaction objects}
\usage{
simplify_reaction(x, ...)

\method{simplify_reaction}{reaction}(x, ...)

\method{remove_zero_counts}{reaction}(x, ...)

\method{remove_zero_counts}{reaction_list}(x, ...)

\method{simplify_reaction}{reaction_list}(x, ...)
}
\arguments{
\item{x}{A reaction object}

\item{...}{ignored}
}
\value{
A reaction object
}
\description{
Simplify reaction objects
}
\examples{
r <- as_reaction("NH3 + H+ + H2O = H3O+ + NH3")
simplify_reaction(r)

r2 <- as_reaction("0NH3 + H+ + H2O = H3O+")
remove_zero_counts(r2)

}
