% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{sort.mol}
\alias{sort.mol}
\title{Sort molecule, reaction objects}
\usage{
\method{sort}{mol}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{An object}

\item{decreasing}{Should the list be sorted}

\item{...}{Passed to \link[base]{order}}
}
\value{
An object with the same class as x
}
\description{
Sorts mol objects by alphabetic character representation, and
reaction objects by decreasing coefficient. This is probably useful
for comparison.
}
\examples{
sort(mol(~H2O, ~Ar, ~H2))
sort(as_reaction("2H2O + 4H+ = 2H2O + 4H+"))

}
