% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{*.reaction}
\alias{*.reaction}
\title{Reaction arithmetic}
\usage{
\method{*}{reaction}(x, y)
}
\arguments{
\item{x}{A reaction object}

\item{y}{A reaction object or operand}
}
\value{
A reaction object
}
\description{
Reaction arithmetic
}
\examples{
r1 <- as_reaction("O2 + 2H2 = 2H2O", log_k = -46.62)
r2 <- as_reaction("H3O+ + OH- = 2H2O", log_k = 14)

-r1
-r2
r1 + r2
r1 - r2
r1 * 2
2 * r1
r1 * -2

}
