% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{as.character.reaction}
\alias{as.character.reaction}
\alias{as.character.reaction_list}
\alias{print.reaction}
\alias{print.reaction_list}
\title{Coerce reactions to a character vector}
\usage{
\method{as.character}{reaction}(x, equals_sign = "=",
  wrap_coeff = identity, wrap_super = identity, wrap_sub = identity,
  mol_sep = " + ", ...)

\method{as.character}{reaction_list}(x, ...)

\method{print}{reaction}(x, ...)

\method{print}{reaction_list}(x, ...)
}
\arguments{
\item{x}{A reaction object}

\item{equals_sign, mol_sep}{An equals/plus sign to use for the reaction}

\item{wrap_coeff}{Wrap coefficients for fancy formatting}

\item{wrap_super, wrap_sub}{Wrap super/subscript for fancy formatting}

\item{...}{Ignored}
}
\value{
A character vector
}
\description{
Coerce reactions to a character vector
}
\examples{
r <- as_reaction("2H2 + O2 = 2H2O")
as.character(r)
print(r)

}
