% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{c.molecule_single}
\alias{c.molecule_single}
\alias{c.mol}
\alias{[.mol}
\title{Combine, subset molecule(s) objects}
\usage{
\method{c}{molecule_single}(...)

\method{c}{mol}(...)

\method{[}{mol}(x, i, ...)
}
\arguments{
\item{...}{Objects to combine}

\item{x}{A mol object}

\item{i}{The index to extract}
}
\value{
A mol object
}
\description{
Combine, subset molecule(s) objects
}
\examples{
c(as_mol(~H2O), as_mol(~`NH3`))
mols <- as_mol(c("H2O", "NH3"))
mols[1]

}
