% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{c.molecule_single}
\alias{c.molecule_single}
\alias{c.molecules}
\alias{[.molecules}
\title{Combine, subset molecule(s) objects}
\usage{
\method{c}{molecule_single}(...)

\method{c}{molecules}(...)

\method{[}{molecules}(x, i, ...)
}
\arguments{
\item{...}{Objects to combine}

\item{x}{A molecules object}

\item{i}{The index to extract}
}
\value{
A molecules object
}
\description{
Combine, subset molecule(s) objects
}
\examples{
c(as_molecules(~H2O), as_molecules(~`NH3`))
mols <- as_molecules(c("H2O", "NH3"))
mols[1]

}
