% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{mass}
\alias{mass}
\alias{mass.default}
\alias{mass.molecule_single}
\alias{mass.mol}
\alias{mass.reaction}
\alias{mass.reaction_list}
\alias{charge}
\alias{charge<-}
\alias{charge<-.molecule_single}
\alias{charge<-.mol}
\alias{charge.default}
\alias{charge.reaction}
\alias{charge.reaction_list}
\alias{charge.molecule_single}
\alias{charge.mol}
\title{Access properties of a molecule object}
\usage{
mass(x)

\method{mass}{default}(x)

\method{mass}{molecule_single}(x)

\method{mass}{mol}(x)

\method{mass}{reaction}(x)

\method{mass}{reaction_list}(x)

charge(x)

charge(x) <- value

\method{charge}{molecule_single}(x) <- value

\method{charge}{mol}(x) <- value

\method{charge}{default}(x)

\method{charge}{reaction}(x)

\method{charge}{reaction_list}(x)

\method{charge}{molecule_single}(x)

\method{charge}{mol}(x)
}
\arguments{
\item{x}{A molecule object or symbol string}

\item{value}{A replacement value}
}
\value{
The mass of the molecule or element in g/mol
}
\description{
Access properties of a molecule object
}
\examples{
mass("H")
m <- as_molecule_single("H2O")
mass(m)

}
