% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{new_reaction}
\alias{new_reaction}
\alias{validate_reaction}
\alias{is_reaction}
\alias{new_reaction_list}
\alias{is_reaction_list}
\alias{validate_reaction_list}
\title{Create, validate reaction objects}
\usage{
new_reaction(x)

validate_reaction(x)

is_reaction(x)

new_reaction_list(x)

is_reaction_list(x)

validate_reaction_list(x)
}
\arguments{
\item{x}{A reaction object}
}
\value{
A reaction object
}
\description{
Create, validate reaction objects
}
\examples{
r <- new_reaction(list(mol = as_mol(c("H2O", "H+", "H3O+")),
                       coefficient = c(1, 1, -1)))
validate_reaction(r)

}
