% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule.R
\name{print.molecule_single}
\alias{print.molecule_single}
\alias{print.mol}
\alias{as.character.molecule_single}
\alias{as.character.mol}
\title{Coerce molecule(s) to character}
\usage{
\method{print}{molecule_single}(x, ...)

\method{print}{mol}(x, ...)

\method{as.character}{molecule_single}(x, ...)

\method{as.character}{mol}(x, ...)
}
\arguments{
\item{x}{A molecule(s) object}

\item{...}{Ignored}
}
\value{
A character vector
}
\description{
Coerce molecule(s) to character
}
\examples{
print(as_molecule_single(~H2O))
print(as_mol(~H2O))
as.character(NA_molecule_)
as.character(mol(NA_molecule_))

}
