% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{reaction}
\alias{reaction}
\alias{reaction_list}
\alias{as_reaction}
\alias{as.reaction}
\alias{as_reaction.reaction}
\alias{as_reaction.mol}
\alias{as_reaction.formula}
\alias{as_reaction.character}
\alias{as_reaction_list}
\alias{as.reaction_list}
\alias{as_reaction_list.reaction_list}
\alias{as_reaction_list.list}
\alias{as_reaction_list.reaction}
\alias{as_reaction_list.character}
\title{Create reaction objects}
\usage{
reaction(lhs, rhs, counts_lhs = 1, counts_rhs = 1, log_k = NA_real_,
  validate = TRUE)

reaction_list(..., validate = TRUE)

as_reaction(x, ...)

as.reaction(x, ...)

\method{as_reaction}{reaction}(x, ...)

\method{as_reaction}{mol}(x, coefficient = rep_len(1, length(x)),
  log_k = NA_real_, validate = TRUE, ...)

\method{as_reaction}{formula}(x, log_k = NA_real_, validate = TRUE,
  ...)

\method{as_reaction}{character}(x, log_k = NA_real_, validate = TRUE,
  ...)

as_reaction_list(x, ...)

as.reaction_list(x, ...)

\method{as_reaction_list}{reaction_list}(x, ...)

\method{as_reaction_list}{list}(x, validate = TRUE, ...)

\method{as_reaction_list}{reaction}(x, validate = TRUE, ...)

\method{as_reaction_list}{character}(x, validate = TRUE, ...)
}
\arguments{
\item{lhs}{The lefthand side}

\item{rhs}{The righthand side}

\item{counts_lhs}{The counts for molecules on the lefthand side}

\item{counts_rhs}{The counts for molecules on the righthand side}

\item{log_k}{The base 10 logarithm of the reaction quotient}

\item{validate}{Flag to validate molecules in the reaction}

\item{...}{Passed to/from methods}

\item{x}{An object to convert to a reaction object}

\item{coefficient}{coefficient corresponding to as_reaction input}
}
\value{
A reaction object
}
\description{
Create reaction objects
}
\examples{
reaction(c("H2O", "H+"), "H3O+", log_k = 14)

}
